PWM from voltage
Do it properly, rather than guessing at a multiple of 2000. We multiply
by max value over max voltage. Also changed LQR and observer poles, but
neither really improves the results.

The state estimates don't quite follow the measurements, which likely is
the issue. Maybe adding in the smaller pendulum and maybe the wire as a
spring would work.